requirements.txt
fastapi
uvicorn
pandas
numpy
pymongo
requests
pytz
jinja2
python-dotenv
python-multipart